requirements.txt
aiodns>=1.1  # aiodns speeds up DNS resolving
aiofile>=2.0.0
aiohttp==3.8.4
cchardet
cython
order_book==0.6.0
pyyaml
requests>=2.18.4
uvloop
websockets>=7.0
yapic.json>=1.6.3
pythclient==0.1.4